+++ rs/execution_environment/src/query_handler/query_cache.rs
@@ -5,7 +5,10 @@ use ic_metrics::MetricsRegistry;
 use ic_query_stats::QueryStatsCollector;
 use ic_replicated_state::ReplicatedState;
 use ic_types::{
-    batch::QueryStats, ingress::WasmResult, messages::Query, Cycles, MemoryDiskBytes, Time, UserId,
+    batch::QueryStats,
+    ingress::WasmResult,
+    messages::{CertificateDelegationFormat, CertificateDelegationMetadata, Query},
+    Cycles, MemoryDiskBytes, Time, UserId,
 };
 use ic_utils_lru_cache::LruCache;
 use prometheus::{Histogram, IntCounter, IntGauge};
@@ -135,6 +138,24 @@ pub(crate) struct EntryKey {
     pub method_name: String,
     /// Receiving canister method payload (argument).
     pub method_payload: Vec<u8>,
+    /// Format of the certificate delegation.
+    pub certificate_delegation_format: Option<CertificateDelegationFormat>,
+}
+
+impl EntryKey {
+    pub fn new(
+        query: &Query,
+        certificate_delegation_metadata: Option<CertificateDelegationMetadata>,
+    ) -> Self {
+        Self {
+            source: query.source.user_id(),
+            receiver: query.receiver,
+            method_name: query.method_name.clone(),
+            method_payload: query.method_payload.clone(),
+            certificate_delegation_format: certificate_delegation_metadata
+                .map(|metadata| metadata.format),
+        }
+    }
 }
 
 impl MemoryDiskBytes for EntryKey {
@@ -144,17 +165,6 @@ impl MemoryDiskBytes for EntryKey {
 
     fn disk_bytes(&self) -> usize {
         0
-    }
-}
-
-impl From<&Query> for EntryKey {
-    fn from(query: &Query) -> Self {
-        Self {
-            source: query.source.user_id(),
-            receiver: query.receiver,
-            method_name: query.method_name.clone(),
-            method_payload: query.method_payload.clone(),
-        }
     }
 }